Expose environment variable for currently running scenario.
In some circumstances, you may want to know the scenario currently being
ran (i.e. one example is generating test results files for each scenario).
This exposes the current scenario name as `EMBER_TRY_CURRENT_SCENARIO`
just before running the command and resets it just before cleaning up.

My personal use case for this is to generate a testem report file for
each scenario, and to do that I need to be able to detect the scenario
in my projects Testem config...