Leverage command return value instead of process.exit.
Works around an upstream issue in ember-cli 2.13 canary versions,
and is quite a bit easier to test.

Also tweaked SIGINT handling to queue up a cancel and avoid additional
work. This is because in ember-cli 2.9+ we use `capture-exit` so that
`process.exit()` does not immediately schedule an exit any longer (so
the semantics that we intended with the prior SIGINT logic were generally
broken already).