Bump ember-cli dependency
This is because 2.8.0 depended on npm 2.15.5 which was run for scenario
npm installation.  During the execution of `./all-commands.sh`, on
certain systems this would involve traversing the circular symlink
`./node_modules/ember-try` many times.  At a minimum this would cause
postinstall for eg spawn to be run repeatedly, but it would sometimes
result in `ELOOP: too many symbolic links`

There is no particular reason for the test suite to rely on such an old
version of npm.