Ensure `bower` is not required when no bower dependencies are present.
When a scenario set includes at least one scenario with `bower: { /*
stuff */ }` **all** scenarios in that scenario set will attempt to run
`bower install` (depsite it not being installed).

With these changes, the bower adapter will still be present but it is
now aware that it should not run `bower install` if a given scenario
does not have bower dependencies/devDependencies.