[feat] Adds support for Yarn/NPM workspaces
This PR adds support for workspaces to allow ember-try to run scenarios
in monorepo based repositories quickly and efficiently. Workspaces are
currently only supported by Yarn, but NPM has indicated that they
[want to add support in the future](https://blog.npmjs.org/post/173239798780/beyond-npm6-the-future-of-the-npm-cli),
and other tools for managing monorepos such as Lerna integrate nicely
with workspaces, so it feels like the best way to add this type of
support.

To enable workspaces, users need to install `ember-cli` and `ember-try`
in the top level monorepo package, add a `config/ember-try.js` file, and
set `useWorkspaces` and `useYarn` to true in that config. This means
that the adapter is limited to using one try config for _all_ of the
packages in the repo. For most monorepo use cases this should be
sufficient, and users can still add ember-try to individual packages if
needed for unique per-package scenarios.

The workspace adapter reuses the NPM adapter for each individual
package. To do this, a new method needed to be exposed to allow depSets
to be applied without running the install, since that needs to be done
_once_ at the top level. An alternative would be to expose this as a
task directly so it could be run as a command in each package, but that
would be significantly more complicated.