Restore dependencies after running scenarios IFF there are multiple scenarios.
When there is only a single scenario to run (e.g. `ember try:one`) we should not be cleaning up as
part of running the scenario itself because in that case we would immediately cleanup _anyways_. If
we forced cleanup to happen during the scenario run, then it would be impossible to have
`--skip-cleanup` work at all.